+++ src/main/java/com/vonage/client/redact/RedactRequest.java
@@ -81,10 +81,6 @@ public class RedactRequest {
         NUMBER_INSIGHTS("number-insight"),
         VERIFY("verify"),
         VERIFY_SDK("verify-sdk"),
-        /**
-         * @deprecated Use {@link Product#MESSAGES}
-         **/
-        @Deprecated MESSAGE("messages"),
         MESSAGES("messages"),
         WORKFLOW("workflow");